Avoid splitting server responses
By default, bufio.Scanner will split data on new lines. This reflected
in the websocket responses as JSON strings get broken into separate
lines, which complicates parsing of such responses.

This introduces a custom bufio.SplitFunc that reads the chunk of data
(up till EOF) and returns the data as is.

Signed-off-by: idoko <okokomichaels@outlook.com>